Refactored code for modal nav buttons
Code was actually fine INSIDE of displayModal function, but needed to be refactored to target the buttons themselves, not the modal container. Slack moderator was a big help explaining what the error meant. Was able to refactor and resolve issue quickly once learning what "cannot destructure property 'name' of 'employees(index)' as it is" error meant